ui: shop: move shop dir
