+++ src/App.js
@@ -13,8 +13,8 @@ import Navbar from './comps/global/Navbar'
 import Home from './comps/main/Home'
 import Login from './comps/main/Login'
 import Profile from './comps/account/Profile'
-import Shop from './comps/account/Shop'
-import ShopItem from './comps/account/ShopItem'
+import Shop from './comps/shop/Shop'
+import ShopItem from './comps/shop/ShopItem'
 import Dashboard from './comps/account/Dashboard'
 import Footer from './comps/global/Footer'
 import Send from './comps/account/Send'
